refactor(ux): :recycle: Switch to Serwist configurator mode for bundler-agnostic SW build
Replace @serwist/turbopack (webpack/bundler-coupled) with @serwist/cli configurator mode.
Serwist now runs as a separate post-build step via `serwist build`, reading a standalone
serwist.config.js that uses serwist.withNextConfig to receive the resolved Next.js config.

Key changes:
- next.config.ts has zero Serwist code — clean separation, swap bundler freely
- serwist.config.js: globIgnores exclude images/, PDFs, and search-index.json, reducing
  the precache from 290 MB (1737 URLs) to 65.8 MB (684 URLs); oversized files skipped
- package.json: predev/build scripts invoke serwist build after next build
- sw.ts: uses /// <reference> directives to scope webworker types to this file only
- tsconfig.json: excludes public/sw.js from TypeScript compilation
- .gitignore: excludes generated public/sw* and public/swe-worker* files
