--- src/app/serwist/[path]/route.ts
@@ -1,13 +1,0 @@
-import { spawnSync } from "node:child_process";
-import { createSerwistRoute } from "@serwist/turbopack";
-
-const revision =
-    spawnSync("git", ["rev-parse", "HEAD"], { encoding: "utf-8" }).stdout?.trim() ??
-    crypto.randomUUID();
-
-export const { dynamic, dynamicParams, revalidate, generateStaticParams, GET } =
-    createSerwistRoute({
-        swSrc: "src/app/sw.ts",
-        additionalPrecacheEntries: [{ url: "/offline", revision }],
-        useNativeEsbuild: true,
-    });
