+++ src/components/sections/pwa/components/serwist-provider.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import { SerwistProvider as BaseSerwistProvider } from "@serwist/turbopack/react";
+import { SerwistProvider as BaseSerwistProvider } from "@serwist/next/react";
 import { ReactNode } from "react";
 
 export function SerwistProvider({ children }: { children: ReactNode }) {
     return (
-        <BaseSerwistProvider swUrl="/serwist/sw.js">
+        <BaseSerwistProvider swUrl="/sw.js">
             {children}
         </BaseSerwistProvider>
     );
