+++ tsconfig.json
@@ -37,6 +37,7 @@
   ],
   "exclude": [
     "node_modules",
-    "agent"
+    "agent",
+    "public/sw.js"
   ]
 }
